requirements.txt
youtube_dl
PyYAML
asyncpg
mutagen
git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py[voice]